+++ uliweb/i18n/__init__.py
@@ -10,7 +10,13 @@ import threading
 import locale
 from uliweb.utils.common import safe_unicode
 
-from lazystr import lazy
+from lazystr import lazy, LazyString
+
+def lazy_ini_convertor(v):
+    from uliweb.utils.pyini import uni_prt
+    return "_(%s)" % uni_prt(v.msg, encoding='utf8')
+
+i18n_ini_convertor = {LazyString:lazy_ini_convertor}
 
 _active_locale = threading.local()